add clippy refactor
